+++ source/Public/Get-ComputerName.ps1
@@ -1,0 +1,37 @@
+<#
+    .SYNOPSIS
+        Returns the computer name cross-plattform.
+
+    .DESCRIPTION
+        The variable `$env:COMPUTERNAME` does not exist cross-platform which
+        hinders development and testing on macOS and Linux. Instead this cmdlet
+        can be used to get the computer name cross-plattform.
+
+    .EXAMPLE
+        Get-ComputerName
+
+        This example returns the computer name cross-plattform.
+#>
+function Get-ComputerName
+{
+    [CmdletBinding()]
+    [OutputType([System.String])]
+    param ()
+
+    $computerName = $null
+
+    if ($IsLinux -or $IsMacOs)
+    {
+        $computerName = hostname
+    }
+    else
+    {
+        <#
+            We could run 'hostname' on Windows too, but $env:COMPUTERNAME
+            is more widely used.
+        #>
+        $computerName = $env:COMPUTERNAME
+    }
+
+    return $computerName
+}